update measure current

+++ MDK-ARM/current_sensor.h
@@ -1,3 +1,3 @@
 #include "main.h"
 
-void ConvertVoltage(uint32_t *voltage_read, float *rawVoltage);
+void ConvertVoltage(uint16_t *voltage_read, float *rawVoltage);
